CruzPlot

Create maps of shipboard DAS data

Details

A utility program oriented to create maps, plot data, and do basic data summaries from data files in the "DAS" format, typically produced by WinCruz.

See also

Author

Sam Woodman sam.woodman@noaa.gov